pyproject.toml
[tool.poetry]
name = "youtube-data"
version = "0.1.0"
description = ""
authors = ["ruoshui-git <ruoshuim@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
selenium = "^3.141.0"
dataclasses-json = "^0.5.2"
typer = "^0.3.2"
google-api-python-client = "^1.12.8"
python-dotenv = "^0.15.0"
SQLAlchemy = "1.4.0b3"
alembic = "^1.5.5"
SQLAlchemy-Utils = "^0.36.8"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"
ipykernel = "^5.4.3"
pandas = "^1.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
